requirements.txt
absl-py==0.9.0
astor==0.8.1
bert-serving-client==1.10.0
bert-serving-server==1.10.0
Click==7.0
Flask==1.1.1
gast==0.2.2
google-pasta==0.1.8
GPUtil==1.4.0
grpcio==1.26.0
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.11.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.18.1
opt-einsum==3.1.0
protobuf==3.11.2
pyzmq==18.1.1
scipy==1.4.1
six==1.14.0
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
Werkzeug==0.16.1
wrapt==1.11.2
